Add a super class to handle backend connection for neptune objects
